+++ src/commands/find.ts
@@ -4,6 +4,7 @@ import { EmacsCommand } from ".";
 import { IEmacsController } from "../emulator";
 import { MessageManager } from "../message";
 import { revealPrimaryActive } from "./helpers/reveal";
+import { WorkspaceConfigCache } from "../workspace-configuration";
 
 export interface SearchState {
   startSelections: readonly vscode.Selection[] | undefined;
@@ -21,16 +22,11 @@ interface FindArgs {
 
 abstract class IsearchCommand extends EmacsCommand {
   protected searchState: SearchState;
-  protected seedSearchStringFromSelection: string | undefined;
 
   public constructor(emacsController: IEmacsController, searchState: SearchState) {
     super(emacsController);
 
     this.searchState = searchState;
-
-    const editorFindConfig = vscode.workspace.getConfiguration("editor.find");
-    const { seedSearchStringFromSelection } = editorFindConfig;
-    this.seedSearchStringFromSelection = seedSearchStringFromSelection;
   }
 
   protected openFindWidget(opts: { isRegex: boolean; replaceString?: string }): Thenable<void> {
@@ -45,7 +41,9 @@ abstract class IsearchCommand extends EmacsCommand {
       preserveCase: false,
     };
 
-    if (this.seedSearchStringFromSelection === "never") {
+    const { seedSearchStringFromSelection } = WorkspaceConfigCache.get("editor.find");
+
+    if (seedSearchStringFromSelection === "never") {
       return vscode.commands.executeCommand("editor.actions.findWithArgs", findArgs);
     }