+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the "emacs-mcx" extension will be documented in this file
 
 ### Fix
 - Dispose unused resources, #1307, #1309.
+- Find methods automatically reload the `editor.find.seedSearchStringFromSelection` config when changed, #1310.
 
 ## [0.43.1] - 2022-04-29
 ### Fix